requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.118
botocore==1.29.118
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
